Add (experimental) helm-specific makefile targets.
Most basic usage will allow you invoke helm via makefile regardless of which helm version you are using. A shim that relies on the result of `helm_is_v2` allows this.

The rest is experimentally written to investigate the relationship between other KUBE_NAMESPACE, HELM_CHART, HELM_RELEASE and the deployed pods/charts, e.g.

- TANGO_HOST is resolved dynamically
- HELM_RELEASE allowed to be set by helm and then inferred later by invoking `make helm_ls`

All deployments in the pipeline are still done by `kubectl apply` instead of `helm install`. We hope to change this when we understand a clear way to do so.