Added basic support for JIRA integration. ST-299
This has added a .feature file exported from JIRA. It also adds a
post-test phase to the CI pipeline to export the results to JIRA. The
results may not be generated properly yet. The associated CI tokens have
been added. TODO: export the test results in a sensible way to the CI
pipeline.