SAR-158 Marked DishMaster tests with xfail
The current tmc-prototype docker image used has missing attribute(s) and command(s) on the DishMaster simulator, which are required as described by the SKA Tango Guidelines. However, at the moment it won't be possible to make a new release of the tmc-prototype docker image as there is still ongoing work. The xfail mark will be removed on the next release of the tmc-mid helm chart.