Add new skamid-emulated umbrella chart
This is a verbatim copy of the skamid umbrella chart, except that
instead of depending on the mid-cbf chart to start the Mid.CBF subsystem
it depends on the mid-cbf-emulated chart. The latter installs a device
server for a simple, dummy CbfMaster device, plus three CbfSubarray
device servers that use the cbf-sdp-emulator sender to data over to SDP.

This work is part of YAN-688.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>